refactor(service): move domain-specific exceptions
