+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/documentation/usecase/ApiCreateDocumentationPageUsecase.java
@@ -20,12 +20,9 @@ import io.gravitee.apim.core.documentation.crud_service.PageCrudService;
 import io.gravitee.apim.core.documentation.domain_service.ApiDocumentationDomainService;
 import io.gravitee.apim.core.documentation.domain_service.CreateApiDocumentationDomainService;
 import io.gravitee.apim.core.documentation.domain_service.HomepageDomainService;
+import io.gravitee.apim.core.documentation.exception.InvalidPageParentException;
 import io.gravitee.apim.core.documentation.model.Page;
-import io.gravitee.apim.core.exception.InvalidPageParentException;
-import io.gravitee.apim.core.sanitizer.HtmlSanitizer;
-import io.gravitee.apim.core.sanitizer.SanitizeResult;
 import io.gravitee.rest.api.service.common.UuidString;
-import io.gravitee.rest.api.service.exceptions.PageContentUnsafeException;
 import java.util.Date;
 import java.util.Objects;
 import lombok.Builder;
